backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
pydantic-settings
httpx
SQLAlchemy
passlib[bcrypt]
python-jose[jwt]
alembic
python-json-logger
psycopg2-binary